requirements.txt
Flask==2.3.2
gunicorn==20.1.0
firebase-admin
google-generativeai
python-dotenv
autopep8
